# front/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/api","description":"Generated server url"}],"paths":{"/devices/{deviceId}":{"put":{"tags":["device-controller"],"operationId":"update","parameters":[{"name":"deviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}},"delete":{"tags":["device-controller"],"operationId":"delete","parameters":[{"name":"deviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/compartments/update":{"put":{"tags":["compartment-controller"],"operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompartmentCreateAndUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/pills/{deviceId}/tray-emptied":{"post":{"tags":["pill-controller"],"operationId":"confirmTrayEmptied","parameters":[{"name":"deviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/pills/{deviceId}/not-taken":{"post":{"tags":["pill-controller"],"operationId":"confirmNotTaken","parameters":[{"name":"deviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PillNotTakenRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/pills/{deviceId}/drop":{"post":{"tags":["pill-controller"],"operationId":"confirmDrop","parameters":[{"name":"deviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PillDropRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/devices/register":{"post":{"tags":["device-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceRegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/devices/connectForManager":{"post":{"tags":["device-controller"],"operationId":"connectDeviceAsAManager","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceRegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/devices/caregiver/requests/{approvalId}/reject":{"post":{"tags":["device-controller"],"operationId":"rejectRequest","parameters":[{"name":"approvalId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/devices/caregiver/requests/{approvalId}/approve":{"post":{"tags":["device-controller"],"operationId":"approveRequest","parameters":[{"name":"approvalId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/devices/caregiver/request":{"post":{"tags":["device-controller"],"operationId":"caregiverRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeviceCaregiverRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/compartments":{"post":{"tags":["compartment-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompartmentCreateAndUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/1.0/logout":{"post":{"tags":["auth-controller"],"operationId":"handleLogout","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}},{"name":"Trend-Auth-Token","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GenericMessage"}}}}}}},"/1.0/login":{"post":{"tags":["auth-controller"],"operationId":"handleAuthentication","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/v1/logs/daily":{"get":{"tags":["activity-log-controller"],"operationId":"getDailyLogs","parameters":[{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivityLogResponse"}}}}}}}},"/devices/{deviceId}/caregiver/requests":{"get":{"tags":["device-controller"],"operationId":"getPendingRequests","parameters":[{"name":"deviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPendingApproval"}}}}}}},"/devices/my":{"get":{"tags":["device-controller"],"operationId":"myDevices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListDeviceResponse"}}}}}}},"/devices/my-device-id":{"get":{"tags":["device-controller"],"operationId":"getMyDeviceId","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLong"}}}}}}},"/compartments/device/{deviceId}":{"get":{"tags":["compartment-controller"],"operationId":"list","parameters":[{"name":"deviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListCompartmentDto"}}}}}}},"/compartments/device/{deviceId}/idx/{idx}":{"delete":{"tags":["compartment-controller"],"operationId":"deleteByDeviceAndIdx","parameters":[{"name":"deviceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idx","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}}},"components":{"schemas":{"DeviceUpdateRequest":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"}}},"ApiResponseVoid":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object"},"path":{"type":"string"},"validationErrors":{"type":"object","additionalProperties":{"type":"string"}}}},"CompartmentCreateAndUpdateRequest":{"required":["deviceId","medicine"],"type":"object","properties":{"deviceId":{"type":"integer","format":"int64"},"idx":{"maximum":4,"minimum":1,"type":"integer","format":"int32"},"medicine":{"$ref":"#/components/schemas/Medicine"},"quantity":{"type":"integer","format":"int32"},"scheduleList":{"maxItems":2147483647,"minItems":1,"type":"array","items":{"type":"string","format":"date-time"}}}},"Medicine":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"dosage":{"type":"string"},"owner":{"$ref":"#/components/schemas/User"},"deleted":{"type":"boolean"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN","MANAGER"]},"status":{"type":"string","enum":["PENDING","ACTIVE","PAUSED","DISABLED","INACTIVE"]},"createdAt":{"type":"string","format":"date-time"},"deleted":{"type":"boolean"},"deletedAt":{"type":"string","format":"date-time"}}},"PillNotTakenRequest":{"required":["pillInstanceId"],"type":"object","properties":{"pillInstanceId":{"type":"integer","format":"int64"}}},"PillDropRequest":{"required":["compartmentIdx","pillInstanceId"],"type":"object","properties":{"pillInstanceId":{"type":"integer","format":"int64"},"compartmentIdx":{"type":"integer","format":"int32"}}},"DeviceRegisterRequest":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"}}},"DeviceCaregiverRequest":{"type":"object","properties":{"name":{"type":"string"}}},"GenericMessage":{"type":"object","properties":{"message":{"type":"string"}}},"Credentials":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":0,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"},"token":{"$ref":"#/components/schemas/Token"},"tokenType":{"type":"string"},"deviceId":{"type":"integer","format":"int64"}}},"Token":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"prefix":{"type":"string"},"token":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN","MANAGER"]},"status":{"type":"string","enum":["PENDING","ACTIVE","PAUSED","DISABLED","INACTIVE"]},"active":{"type":"boolean"}}},"ActivityLogResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"timestamp":{"type":"string","format":"date-time"},"username":{"type":"string"},"actionType":{"type":"string","enum":["USER_LOGIN","USER_LOGOUT","USER_REGISTERED","USER_PASSWORD_RESET","USER_UPDATED","SYSTEM_ERROR","API_CALL","DATABASE_OPERATION"]},"actionTypeDescription":{"type":"string"},"description":{"type":"string"},"details":{"type":"string"}}},"ApiResponseListPendingApproval":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PendingApproval"}},"path":{"type":"string"},"validationErrors":{"type":"object","additionalProperties":{"type":"string"}}}},"Device":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"password":{"type":"string"},"owner":{"$ref":"#/components/schemas/User"},"viewers":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/User"}}}},"PendingApproval":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"requestDate":{"type":"string","format":"date"},"type":{"type":"string","enum":["ROLE","DEVICE_VIEWER"]},"device":{"$ref":"#/components/schemas/Device"}}},"ApiResponseListDeviceResponse":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DeviceResponse"}},"path":{"type":"string"},"validationErrors":{"type":"object","additionalProperties":{"type":"string"}}}},"DeviceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"manager":{"type":"boolean"}}},"ApiResponseLong":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"integer","format":"int64"},"path":{"type":"string"},"validationErrors":{"type":"object","additionalProperties":{"type":"string"}}}},"ApiResponseListCompartmentDto":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CompartmentDto"}},"path":{"type":"string"},"validationErrors":{"type":"object","additionalProperties":{"type":"string"}}}},"CompartmentDto":{"type":"object","properties":{"compartmentId":{"type":"integer","format":"int64"},"idx":{"type":"integer","format":"int32"},"medicineName":{"type":"string"},"dosage":{"type":"string"},"remaining":{"type":"integer","format":"int64"},"dates":{"type":"array","items":{"type":"string","format":"date-time"}}}}}}}